[image output] Initialize pixels of partial tile conversion buffer.
When writing a partial tile, the unused pixels still go through float
conversion.

This means, that floating point operations are done on uninitialized
data.

This can easily lead to NaN and to floating point exceptions, if those
were to be enabled.

This change will set the shared buffer used for all partial tiles to
all zero pixels.

Tested by running valgrind before and after the change

FIXES: #4461

Signed-off-by: Bram Stolk <b.stolk@gmail.com>