ci: Streamline the old MacOS-12 CI test
It's an old flavor not well supported by homebrew any longer, so
what's happening is that most of the dependency packages we ask to
install have to build from source, which means it takes a long long
time.

By disabling OpenCV and Qt homebrew installation for it, we can make
it fast again. Neither the minimal OpenCV interop, nor osltoy (the
only thing that needs Qt) are exercised by the automated testsuite
anyway, so I don't think there's any loss of testing functionality
in practice.

Signed-off-by: Larry Gritz <lg@larrygritz.com>