ci: Remove old MacOS-12 CI test
It's an old flavor not well supported by homebrew any longer, so
what's happening is that most of the dependency packages we ask to
install have to build from source, which means it takes a long long
time. Just get rid of it, we have two other Mac test cases -- a newer
MacOS 13 Intel one, and a MacOS 14 ARM one. That's enough coverage.

Signed-off-by: Larry Gritz <lg@larrygritz.com>
Signed-off-by: Zach Lewis <zachcanbereached@gmail.com>