ci(wheels): disable CIBW repair-wheel step + remove tasks.py
Not only do we remove the custom repairwheel step *entirely*, we want to make sure the default repairwheel step does not run at all.

For some reason, `repairwheel`, `autditwheel`, etc. have trouble recognizing that the wheel first created by the build system is fine as-is, and doesn't need repairing. In a best-case scenario, the repair-wheel step ends up creating and packaging copies of OpenImageIO_Util and OpenColorIO.

We may need to reintroduce a custom repair wheel step if we cannot insulate the build from other dependencies on runner images picked up by OIIO.

Signed-off-by: Zach Lewis <zachcanbereached@gmail.com>