fix: fix incorrect vertical alignment in text render
The y argument in OIIO::ImageBufAlgo::render_text() should represent the top and bottom edge of the text if vertical alignment is set to TextAlignY::Top and TextAlignY::Bottom. However currently the rendered text is off. The current code adjusts y to point to the top of the text box before rendering the text, but it has to be adjusted to point to the baseline of the text.
There is a current test case, testing vertical alignment, but the reference image seems wrong and has now been updated.

Signed-off-by: peter.horvath <peter.horvath@autodesk.com>