fix(IBA): fix incorrect vertical alignment in text render (#4500)
The `y` argument in `OIIO::ImageBufAlgo::render_text()` should represent
the top and bottom edge of the text if vertical alignment is set to
`TextAlignY::Top` and `TextAlignY::Bottom`. However currently the
rendered text is off. The current code adjusts y to point to the top of
the text box before rendering the text, but it has to be adjusted to
point to the baseline of the text.

Tests

There is a existing test case, testing vertical alignment, but the
reference image seems wrong and has now been updated. See the "Topleft"
and "Rightbot" texts.

| Current    | Fixed |
| -------- | ------- |
|
![image](https://github.com/user-attachments/assets/e9e135bf-21d1-43a9-ace9-05b0e99af084)
|
![image](https://github.com/user-attachments/assets/970dee6a-5337-4382-bcc2-16f474bba160)
|

Note: The testsuite contains a reference image for freetype v2.4.11
(aligned-freetype2.4.11.tif). I don't have that version built, and
haven't been able to update the image. I'm wondering if this is still
used or could be removed. I guess CI will tell.

---------

Signed-off-by: peter.horvath <peter.horvath@autodesk.com>