ci: Retire deprecated macos12 runner, try beta macos15
The GHA macos-12 runner is deprecated and soon to be retired. It's
already flaky and they're doing purposeful outages to remind people
(the cloud equivalent of turning the lights on and off right before
the bar closes).

We still want to test earlier MacOS-12 compatibility to match VFX
Reference Platform, so use the macos-13 runner as the minmum but set
MACOSX_DEPLOYMENT_TARGET=12.0 to test that configurations.

Also, the MacOS-15 runner is in beta, so add that to the mix. (Along
the way, that also verifies that we run fine against python 3.13.)

Signed-off-by: Larry Gritz <lg@larrygritz.com>