deps(OCIO): Try to find OCIO dependencies in its own build directory
If we let the build system build OCIO, its build directory should persist for the duration of the parent OIIO build. Try to find the static dependencies that OCIO builds for itself in <OCIO-build-dir>/ext/dist; and if that fails, try to build and install those same static libraries ourselves.

Signed-off-by: Zach Lewis <zachcanbereached@gmail.com>