build: Improve finding locally built PNG
* Adjustments to lock down a better ordering of include directories.
* Always prefer a static library png when available, and when building
  local PNG, for sure only build the static version.
* Give the symbols an "oiio" prefix so they can't link against symbols
  in the wrong library.
* Give a bunch of extra hints about the directory locations and
  config location of the local build to make extra sure it's the
  one found and chosen.

This is still not perfect. It's plagued by one remaining nasty
problem: If there is a too-old version of libpng installed in a system
area, although we will diligently build newer local copy, it's
extremely hard to ensure that it doesn't accidentally get the headers
corresponding to the old version, because OTHER dependencies may be
correctly found in the system area and therefore put it in the include
path where it will be found rather than the one we intend. I don't
know how to fix this robustly.

For now, it remains the case that the dependency self-build works very
well if the dependency is missing, but can be hit-or-miss if there are
headers for a too-old version installed in the same area as other
dependency headers we use, since it's very hard to control which
end up earlier in the include path.

But at least for now, this setup works fine for all of our CI cases.

Signed-off-by: Larry Gritz <lg@larrygritz.com>