fix(exr): allow an empty "name" metadata to be read
When the exr "name" metadata (used for the names of subimages/parts)
existed but was the empty string, we were treating it like it didn't
exist (and upon write substituted our own "subimageXY" name, as we are
wont to do for unnamed parts). But now we've seen multi-part exrs
where the name is "" (exists, but empty string) and we want a
read-write cycle to preserve that. Even though it's not a very
sensible or helpful part name, it's technically legal, so we should
honor it.

Signed-off-by: Larry Gritz <lg@larrygritz.com>