build: Add hardening options
Get rid of the FORTIFY options, which were too specific to one
compiler feature option.  Add OIIO_HARDENING that takes a level
meaning:

- 0 : do nothing, not recommended.
- 1 : enable features that have no (or nearly no) performance impact,
      recommended default for optimized, shipping code.
- 2 : enable features that trade off performance for security, recommended
      for debugging or deploying in security-sensitive environments.
- 3 : enable features that have a significant performance impact, only
      recommended for debugging.

Default to 1 for optimized builds, 3 for debug builds (so will be
thoroughly tested by our sanitizer and other CI tests). Users that
have more stringent security requirements may choose to build with 2
even for shipping code (they should benchmark to see if that is
acceptable to them).

These levels turn on a variety of compiler options that are
recommended for additional safety. We will add more as they are
developed.

There are also some warning suppressions that needed to be added to
code in a few areas where it was unavoidable to use some constructs
that trigger the elevated safety checks.

Signed-off-by: Larry Gritz <lg@larrygritz.com>