ci: refactor using a single steps workflow
The logic for the GHA steps was repeated across multiple workflows and
platforms. Consolidate it all into a common "build-steps" workflow,
and have all the other workflows reference that when possible. This
should hopefully make maintenance of the CI scripts easier in the
future, with less redundancy of the step logic.

The only major place where we deviate is that we have a separeately
specified set of steps for the "old" VFX Platform 2021 and 2022 based
tests (because we need to specify different container arguments to
it).

Signed-off-by: Larry Gritz <lg@larrygritz.com>