api: Add global attribute `imageinput:strict`
We don't do anything with this at the present time, but this PR
reserves and documents this attribute for future use.

The intent is to be able to set whether we want image readers to try
being as tolerant as possible when reading a file with flaws (press on
and see if the rest of the file is ok?), or be more conservative and
abandon reading any file as soon as a corruption or invalid data is
encountered (because that might be a clue that the file is arbitrarily
corrupted or even maliciously constructed).

I documented it as defaulting to 0 (err on the side of being
permissive of bad input), with high-security applications being
responsible for setting it to 1. But it's open for debate of people
think that a better default is to be strict and let applications who
want to be more tolerant be responsible for accepting the risk and
switching the mode.

Signed-off-by: Larry Gritz <lg@larrygritz.com>