fix: IPTC fields have length limits
It escaped our notice before that the IPTC spec dictates length limits
for many fields. Getting this wrong can confuse some software,
including Photoshop apparently.

This patch enforces length limits wherever we could figure them out
from the IPTC spec -- it will simply truncate those strings that are
too long before writing them to a binary IPTC tag.

For this reason, we also are ending the practice of automatically
translating several IPTC fields to and from what we figured were the
equivalent generic metadata names. This was perhaps a dubious practice
to begin with, but now the enforcement of length limits for IPTC (but
not the generic metadata) makes it even more frought. There is
probably nobody depending on this behavior (and maybe few OIIO users
depending on IPTC support at all?), so now anybody purposely using
IPTC metadata is fully responsible for setting it and dealing with any
issues of whether it's "out of sync" with any other metadata that OIIO
stores in or reads from a file.

Fixes 4342

Signed-off-by: Larry Gritz <lg@larrygritz.com>