ci: Allow special branch names to prune jobs
Quite often I find myself debugging the CI for just one platform.  It
makes for much longer iteration for a huge CI job that reruns lots of
tests that won't be affected by the changes I'm experimenting with.
Sometime I make a commit that places `if: 0` in strategic places to
exclude irrelevant batches of tests. But this is a pain to add, and
has to be removed again to re-enable the tests prior to turning the
experiment into a proper PR.

This PR, then, adds a test that lets me do the same just with special
branch names, rather than editing ci.yml. Any branch you push
containing the substring "windows-only" will not run any Linux or
MacOS tests. And the analogous behavior is added for "macos-only" and
"linux-only".

This shouldn't affect any CI jobs in which people are not purposely
using those special branch names. You should NOT use these special
names for branches from which you submit PRs, since of course we want
every PR to validate on all CI combinations. But it is very helpful
for topic branches on your own fork as you are developing and
iterating over many CI runs.

Signed-off-by: Larry Gritz <lg@larrygritz.com>