fix(build): Fix recently broken rpath setting
As part of the big python wheel merge, compiler.cmake got changed in
a subtle way in how it sets CMAKE_INSTALL_RPATH.

This broken iv for me on a Mac with dynamic libraries. Strangely (and
I'm too lazy to chase down exactly why), oiiotool seems fine.  So this
was missed entirely by the CI, which never runs 'iv', since it's an
interactive app. (Note to self: we should find some way to at least
verify that it can execute as part of the testsuite.)

Anyway, I believe that this change adds the install lib area back to
the path, and it does seem to repair my ability to run iv on my end.

Signed-off-by: Larry Gritz <lg@larrygritz.com>