ci: Improved clang-format CI task
Old: Do a build with target 'clang-format', relying on our build
system.  But that means that we need to do the cmake config step, and
need a bunch of dependencies set up so that the config doesn't fail.

New: Make a new run-clang-format.bash utility to issue the commands
for the right files. No "build" is requred, and the only dependencies
we need are the minimum it takes to run clang-format itself.

These changes reduce the approximate total elapsed time of a CI
clang-format job from nearly 7:00 to 0:45.

Also, I noticed that we never had been clang-format testing the .cpp
and .h files in the testsuite itself. So enable that now, and commit
the small amount of reformatting that it identified.

Signed-off-by: Larry Gritz <lg@larrygritz.com>