api(span.h): Only allow span-from-array if array for compatible size
From an array of known size, we can implicitly construct a span of
dynamic length, or of fixed length equal to the array length.  Use
template enable_if trick to restrict to these cases.

Previously, the constructor would have allowed a fixed-length span to
be constructted from an array of differing length, which should
obviously have been disallowed.

Signed-off-by: Larry Gritz <lg@larrygritz.com>